spx_trading_system/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
ta-lib>=0.4.25
pandas-ta>=0.3.14b0

# Backtesting framework
backtrader>=1.9.76.123

# Data handling and APIs
requests>=2.31.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0
seaborn>=0.12.0

# Database
sqlalchemy>=2.0.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
scipy>=1.10.0

# Development
pytest>=7.4.0
black>=23.3.0
flake8>=6.0.0